installer/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "amem-installer"
version = "0.1.0"
description = "One-command installer for AgenticMemory — connects all AI tools to shared memory"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Omoshola Owolabi" },
]
keywords = ["ai", "agent", "memory", "installer", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]

# Only stdlib + PyYAML
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
amem-install = "amem_installer.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["amem_installer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
